done : 1. Deposit 페이지 REMITPAYDE / REMITPAYCO 에 따른 input 조건 변경 2. Deposit, Account 페이지 에러 처리 todo : 1. MediaTracking 페이지 post api 연동 2. OtherWork 페이지 post api 연동 및 CUD 기능 구현 3. MediaTracking 페이지 get api 응답 데이터 적용 수정하기

+++ src/component/Navbar/HeaderNavbar.tsx
@@ -25,16 +25,18 @@ function HeaderNavbar({ routes }: { routes: any[] }) {
 
   // 로그아웃 버튼 클릭 핸들러
   const handleLogout = async () => {
-    try {
-      const response = await axios.delete("/token", { withCredentials: true });
-      if (response.status === 200 || response.status === 204) {
-        // 성공적으로 응답 받으면 로컬 스토리지나 쿠키 비우기
-        localStorage.removeItem("token");
-        navigate("/"); // 로그인 페이지로 이동
+    if(window.confirm("로그아웃 하시겠습니까?")){
+      try {
+        const response = await axios.delete("/token", { withCredentials: true });
+        if (response.status === 200 || response.status === 204) {
+          // 성공적으로 응답 받으면 로컬 스토리지나 쿠키 비우기
+          localStorage.removeItem("token");
+          navigate("/"); // 로그인 페이지로 이동
+        }
+      } catch (err) {
+        console.error("로그아웃 중 오류가 발생했습니다. :", err);
+        alert("로그아웃에 실패했습니다. 다시 시도해주세요.");
       }
-    } catch (err) {
-      console.error("로그아웃 중 오류가 발생했습니다. :", err);
-      alert("로그아웃에 실패했습니다. 다시 시도해주세요.");
     }
   };
